backend/requirements.txt
Flask>=3.0,<4
flask-cors>=4.0,<5
Flask-SQLAlchemy>=3.1,<4
requests>=2.31,<3
Werkzeug>=3.0,<4
bcrypt>=4.1,<5
openpyxl>=3.1,<4
